refactor game Extensions
