+++ WebApi/InWords.WebApi/InWords.Data/Enums/GameTags.cs
@@ -4,6 +4,7 @@
     {
         Official,
         CustomLevelsHistory,
+        CustomPublic,
         Favorite
     }
 }
