+++ WebApi/InWords.WebApi/InWords.Data/InWordsDataContextGame.cs
@@ -1,4 +1,5 @@
-﻿using InWords.Data.Creations.GameBox;
+﻿using InWords.Data.Creations;
+using InWords.Data.Creations.GameBox;
 using Microsoft.EntityFrameworkCore;
 
 namespace InWords.Data
@@ -6,6 +7,7 @@ namespace InWords.Data
     public partial class InWordsDataContext
     {
         public DbSet<GameLevel> GameLevels { get; set; }
+        public DbSet<GameTag> GameTags { get; set; }
 
         public DbSet<GameLevelWord> GameLevelWords { get; set; }
