+++ WebApi/InWords.WebApi/InWords.WebApi/Services/GameService/Models/LevelMetric.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace InWords.WebApi.Services.GameService.Models
+{
+    public class LevelMetric
+    {
+
+    }
+}
